Refactor Offer, OfferPayload
Offer now using OfferPayload with FeeTxOfferPayload as a special case

Add api of common Offer characteristics to OfferPayload.
FeeTxOfferPayload
need to keep the variables that are part of the json contract to work
properly
as they are ordered and the contract hash used in the opreturn of the
deposit tx.